requirements.txt
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0